fix: handle ambiguous calendar schedules
